Ajout de la pagination

+++ src/app/views/favorites/favorites.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
+import { PageEvent } from '@angular/material/paginator';
 import { Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { Favorite } from 'src/app/models/favorite.model';
@@ -17,6 +18,7 @@ export class FavoritesComponent implements OnInit, OnDestroy {
   favorites: Array<Favorite>;
   favoritesSubscription: Subscription;
 
+  pageSlice: Array<Favorite>;
   loading: boolean;
   urlServer = environment.urlServer + 'properties/uploads/';
 
@@ -27,6 +29,7 @@ export class FavoritesComponent implements OnInit, OnDestroy {
     this.favoritesSubscription = this.propertyService.getFavorites().subscribe(
       (favorites: Array<Favorite>) => {
         this.favorites = favorites;
+        this.pageSlice = this.favorites.slice(0, 5);
         this.loading = false;
       },
       (error) => {
@@ -37,18 +40,30 @@ export class FavoritesComponent implements OnInit, OnDestroy {
   }
 
   onDeleteFavorite(favoriteId: number): void {
-    this.loading = true;
-    this.favoritesSubscription = this.propertyService.deleteFavorite(favoriteId).subscribe(
-      (favorites: Array<Favorite>) => {
-        this.favorites = favorites;
-        this.snackbarService.successSnackbar('Hébergement supprimé avec succès de vos favoris.');
-        this.loading = false;
-      },
-      (error) => {
-        console.log('Erreur ! : ' + JSON.stringify(error.error));
-        this.loading = false;
-      }
-    );
+    if (confirm('Êtes-vous sûr de vouloir supprimer cet hébergement de vos favoris ?')) {
+      this.loading = true;
+      this.favoritesSubscription = this.propertyService.deleteFavorite(favoriteId).subscribe(
+        (favorites: Array<Favorite>) => {
+          this.favorites = favorites;
+          this.pageSlice = this.favorites.slice(0, 5);
+          this.snackbarService.successSnackbar('Hébergement supprimé avec succès de vos favoris.');
+          this.loading = false;
+        },
+        (error) => {
+          console.log('Erreur ! : ' + JSON.stringify(error.error));
+          this.loading = false;
+        }
+      );
+    }
+  }
+
+  OnPageChange(event: PageEvent): void {
+    const startIndex = event.pageIndex * event.pageSize;
+    let endIndex = startIndex + event.pageSize;
+    if (endIndex > this.favorites.length) {
+      endIndex = this.favorites.length;
+    }
+    this.pageSlice = this.favorites.slice(startIndex, endIndex);
   }
 
   ngOnDestroy(): void {
